status page

+++ tasks.py
@@ -1,10 +1,13 @@
+#!/usr/bin/env python
+
 from celery import Celery
 import urllib, json, os, jinja2
 from ansible.playbook import PlayBook
 from ansible import callbacks
 from ansible import utils
+from config import CREDS
 
-app = Celery('tasks', backend='amqp://guest@localhost//', broker='amqp://guest@localhost//')
+app = Celery('tasks', backend='db+mysql://' + CREDS.get('user') + ':' + CREDS.get('pass') + '@' + CREDS.get('host') + '/' + CREDS.get('db'), broker='amqp://guest@localhost//')
 
 @app.task
 def install_stuff(ip_list, sshpass, pb_name):
@@ -21,6 +24,11 @@ def install_stuff(ip_list, sshpass, pb_name):
         host_list=ip_list,
     )
     results = pb.run()
-    os.remove(pb_name)
+    if results[ip_list[:-1]]['unreachable'] != 0:
+        raise RuntimeError('df//' + ip_list[:-1] + '//' + '1' + '//fd')
 
-    return results
+    os.remove(pb_name)
+    fail = str(results[ip_list[:-1]]['failures'])
+   
+    ans = '//' + ip_list[:-1] + '//' + fail + '//'
+    return ans
